+++ two_tsp/local_search.py
@@ -3,13 +3,14 @@ import time
 from typing import List, Tuple
 from two_tsp.helpers import cycles_cost, delta_vertex, delta_edge
 
+
 def local_steepest_vertices(
-    cycle1: List[int], cycle2: List[int], dm: List[List[int]]
+    cycle1: List[int], cycle2: List[int], dm: List[List[float]]
 ) -> Tuple[List[int], List[int]]:
     c1 = cycle1[:]
     c2 = cycle2[:]
     while True:
-        best_delta = 0
+        best_delta = 0.0
         best_move = None
         for i in range(len(c1)):
             for j in range(i + 1, len(c1)):
@@ -32,13 +33,14 @@ def local_steepest_vertices(
             c2[i], c2[j] = c2[j], c2[i]
     return c1, c2
 
+
 def local_steepest_edges(
-    cycle1: List[int], cycle2: List[int], dm: List[List[int]]
+    cycle1: List[int], cycle2: List[int], dm: List[List[float]]
 ) -> Tuple[List[int], List[int]]:
     c1 = cycle1[:]
     c2 = cycle2[:]
     while True:
-        best_delta = 0
+        best_delta = 0.0
         best_move = None
         for i in range(len(c1)):
             for j in range(i + 2, len(c1)):
@@ -61,8 +63,9 @@ def local_steepest_edges(
             c2 = c2[: i + 1] + list(reversed(c2[i + 1 : j + 1])) + c2[j + 1 :]
     return c1, c2
 
+
 def local_greedy_vertices(
-    cycle1: List[int], cycle2: List[int], dm: List[List[int]]
+    cycle1: List[int], cycle2: List[int], dm: List[List[float]]
 ) -> Tuple[List[int], List[int]]:
     c1 = cycle1[:]
     c2 = cycle2[:]
@@ -80,20 +83,21 @@ def local_greedy_vertices(
         for which, i, j in pairs:
             if which == 1:
                 d = delta_vertex(dm, c1, i, j)
-                if d < 0:
+                if d < 0.0:
                     c1[i], c1[j] = c1[j], c1[i]
                     improved = True
                     break
             else:
                 d = delta_vertex(dm, c2, i, j)
-                if d < 0:
+                if d < 0.0:
                     c2[i], c2[j] = c2[j], c2[i]
                     improved = True
                     break
     return c1, c2
 
+
 def local_greedy_edges(
-    cycle1: List[int], cycle2: List[int], dm: List[List[int]]
+    cycle1: List[int], cycle2: List[int], dm: List[List[float]]
 ) -> Tuple[List[int], List[int]]:
     c1 = cycle1[:]
     c2 = cycle2[:]
@@ -111,20 +115,21 @@ def local_greedy_edges(
         for which, i, j in pairs:
             if which == 1:
                 d = delta_edge(dm, c1, i, j)
-                if d < 0:
+                if d < 0.0:
                     c1 = c1[: i + 1] + list(reversed(c1[i + 1 : j + 1])) + c1[j + 1 :]
                     improved = True
                     break
             else:
                 d = delta_edge(dm, c2, i, j)
-                if d < 0:
+                if d < 0.0:
                     c2 = c2[: i + 1] + list(reversed(c2[i + 1 : j + 1])) + c2[j + 1 :]
                     improved = True
                     break
     return c1, c2
 
+
 def random_walk(
-    cycle1: List[int], cycle2: List[int], dm: List[List[int]], time_limit: float = 1.0
+    cycle1: List[int], cycle2: List[int], dm: List[List[float]], time_limit: float = 1.0
 ) -> Tuple[List[int], List[int]]:
     start = time.time()
     c1 = cycle1[:]
